+++ source/Playnite.DesktopApp/DesktopCollectionView.cs
@@ -132,7 +132,7 @@ namespace Playnite.DesktopApp
 
         private void Settings_PropertyChanged(object sender, PropertyChangedEventArgs e)
         {
-            if (e.PropertyName == nameof(PlayniteSettings.NameFilterSearchWithAcronyms))
+            if (e.PropertyName == nameof(PlayniteSettings.FuzzyMatchingInNameFilter))
             {
                 Logger.Debug("Refreshing collection view filter.");
                 CollectionView.Refresh();