+++ source/Playnite/Converters/AfterGameLaunchOptionToStringConverter.cs
@@ -15,6 +15,11 @@ namespace Playnite.Converters
     {
         public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
+            if (value == null)
+            {
+                return string.Empty;
+            }
+
             var source = (AfterLaunchOptions)value;
             switch (source)
             {