+++ source/Playnite/Database/InMemoryGameDatabase.cs
@@ -200,7 +200,7 @@ namespace Playnite.Database
             throw new NotImplementedException();
         }
 
-        public bool GetGameMatchesFilter(Game game, FilterPresetSettings filterSettings, bool nameFilterSearchWithAcronyms)
+        public bool GetGameMatchesFilter(Game game, FilterPresetSettings filterSettings, bool useFuzzyNameMatch)
         {
             return true;
         }
@@ -210,17 +210,17 @@ namespace Playnite.Database
             yield break;
         }
 
-        public IEnumerable<Game> GetFilteredGames(FilterPresetSettings filterSettings, bool nameFilterSearchWithAcronyms)
+        public IEnumerable<Game> GetFilteredGames(FilterPresetSettings filterSettings, bool useFuzzyNameMatch)
         {
             yield break;
         }
 
-        public bool GetGameMatchesFilter(Game game, FilterSettings filterSettings, bool nameFilterSearchWithAcronyms)
+        public bool GetGameMatchesFilter(Game game, FilterSettings filterSettings, bool useFuzzyNameMatch)
         {
             return true;
         }
 
-        public IEnumerable<Game> GetFilteredGames(FilterSettings filterSettings, bool nameFilterSearchWithAcronyms)
+        public IEnumerable<Game> GetFilteredGames(FilterSettings filterSettings, bool useFuzzyNameMatch)
         {
             yield break;
         }
@@ -228,6 +228,11 @@ namespace Playnite.Database
         public bool GetGameMatchesFilter(Game game, FilterPresetSettings filterSettings)
         {
             return true;
+        }
+
+        public HashSet<string> GetImportedExeFiles()
+        {
+            throw new NotImplementedException();
         }
     }
 }