+++ source/Playnite/Localization/LocalizationKeys.cs
@@ -170,7 +170,7 @@ namespace Playnite
         /// </summary>
         public const string GameUninstallRunningError = "LOCGameUninstallRunningError";
         /// <summary>
-        /// Are you sure you want to remove this game?
+        /// Are you sure you want to remove {0}?
         /// </summary>
         public const string GameRemoveAskMessage = "LOCGameRemoveAskMessage";
         /// <summary>
@@ -178,7 +178,7 @@ namespace Playnite
         /// </summary>
         public const string GamesRemoveAskMessage = "LOCGamesRemoveAskMessage";
         /// <summary>
-        /// Are you sure you want to remove this game?
+        /// Are you sure you want to remove {0}?
         /// </summary>
         public const string GameRemoveAskMessageIgnoreOption = "LOCGameRemoveAskMessageIgnoreOption";
         /// <summary>
@@ -1278,6 +1278,10 @@ namespace Playnite
         /// </summary>
         public const string SettingsAfterGameClose = "LOCSettingsAfterGameClose";
         /// <summary>
+        /// Format time played to indicate the number of days played
+        /// </summary>
+        public const string SettingsPlaytimeUseDaysFormatLabel = "LOCSettingsPlaytimeUseDaysFormatLabel";
+        /// <summary>
         /// Dates formats:
         /// </summary>
         public const string SettingsDatesFormatsLabel = "LOCSettingsDatesFormatsLabel";
@@ -1589,6 +1593,14 @@ namespace Playnite
         /// Use when experiencing issues with web views, for example integration authentication dialogs.
         /// </summary>
         public const string SettingsAltWebViewRenderingTooltip = "LOCSettingsAltWebViewRenderingTooltip";
+        /// <summary>
+        /// Partial loading of large game descriptions
+        /// </summary>
+        public const string SettingsPartialDescriptionLoading = "LOCSettingsPartialDescriptionLoading";
+        /// <summary>
+        /// Large descriptions can cause noticeable lag when selecting games.
+        /// </summary>
+        public const string SettingsPartialDescriptionLoadingTooltip = "LOCSettingsPartialDescriptionLoadingTooltip";
         /// <summary>
         /// Metadata Import
         /// </summary>
@@ -2878,6 +2890,10 @@ namespace Playnite
         /// </summary>
         public const string LastPlayed = "LOCLastPlayed";
         /// <summary>
+        /// {0}d {1}h {2}m
+        /// </summary>
+        public const string PlayedDays = "LOCPlayedDays";
+        /// <summary>
         /// {0}h {1}m
         /// </summary>
         public const string PlayedHours = "LOCPlayedHours";
@@ -3186,6 +3202,10 @@ namespace Playnite
         /// </summary>
         public const string StatsTopPlayTime = "LOCStatsTopPlayTime";
         /// <summary>
+        /// Total install size
+        /// </summary>
+        public const string StatsTotalInstallSize = "LOCStatsTotalInstallSize";
+        /// <summary>
         /// Overview
         /// </summary>
         public const string OverviewLabel = "LOCOverviewLabel";
@@ -3418,9 +3438,13 @@ namespace Playnite
         /// </summary>
         public const string SelectActionTitle = "LOCSelectActionTitle";
         /// <summary>
-        /// Tracking mode
+        /// Tracking Mode
         /// </summary>
         public const string ActionTrackingMode = "LOCActionTrackingMode";
+        /// <summary>
+        /// Tracking Path
+        /// </summary>
+        public const string ActionTrackingPath = "LOCActionTrackingPath";
         /// <summary>
         /// Link
         /// </summary>
@@ -3754,6 +3778,10 @@ namespace Playnite
         /// </summary>
         public const string EmulatorCustomProfiles = "LOCEmulatorCustomProfiles";
         /// <summary>
+        /// Custom profile
+        /// </summary>
+        public const string EmulatorCustomProfile = "LOCEmulatorCustomProfile";
+        /// <summary>
         /// Handled by a built-in script
         /// </summary>
         public const string EmulatorFunctionHandledByScript = "LOCEmulatorFunctionHandledByScript";
@@ -3893,6 +3921,14 @@ namespace Playnite
         /// Scan inside archives
         /// </summary>
         public const string EmuImportScanInsideArchives = "LOCEmuImportScanInsideArchives";
+        /// <summary>
+        /// Merge related files
+        /// </summary>
+        public const string EmuMergeRelatedFiles = "LOCEmuMergeRelatedFiles";
+        /// <summary>
+        /// Merge related game files, like individual game discs, under one game entry.
+        /// </summary>
+        public const string EmuMergeRelatedFilesTooltip = "LOCEmuMergeRelatedFilesTooltip";
         /// <summary>
         /// Add scanner
         /// </summary>
@@ -4206,13 +4242,13 @@ namespace Playnite
         /// </summary>
         public const string SearchIncludeCommandsInDefaultTooltip = "LOCSearchIncludeCommandsInDefaultTooltip";
         /// <summary>
-        /// Support search using acronyms for name filter
+        /// Use fuzzy matching in name filter
         /// </summary>
-        public const string NameFilterSearchWithAcronyms = "NameFilterSearchWithAcronyms";
+        public const string NameFilterUseFuzzyMatching = "NameFilterUseFuzzyMatching";
         /// <summary>
-        /// When enabled, search results will include items whose acronym matches the search term.
+        /// When enabled, name filter will match game names the same way as global search.
         /// </summary>
-        public const string NameFilterSearchWithAcronymsTooltip = "NameFilterSearchWithAcronymsTooltip";
+        public const string NameFilterUseFuzzyMatchingTooltip = "NameFilterUseFuzzyMatchingTooltip";
         /// <summary>
         /// Fields to be displayed for game results:
         /// </summary>
@@ -4381,5 +4417,17 @@ namespace Playnite
         /// Game menu
         /// </summary>
         public const string GameMenu = "LOCGameMenu";
+        /// <summary>
+        /// Program folder
+        /// </summary>
+        public const string ProgramFolder = "LOCProgramFolder";
+        /// <summary>
+        /// User data directory
+        /// </summary>
+        public const string UserDataFolder = "LOCUserDataFolder";
+        /// <summary>
+        /// Library file corruption has been detected, Playnite will now shutdown.
+        /// </summary>
+        public const string DBCorruptionCrashMessage = "LOCDBCorruptionCrashMessage";
     }
 }